Move CustomAssertion class to its own file
